Don't do custom buffer copy unnecessarily
Only normalize the image buffer into a contiguous block of memory if the
stride > width. If stride is the same as width, that means the image
returned from the screenshotClient is contiguous.

This is vendor specific, e.g. on the S4, we need to normalize the image
data, but on the Nexus 4, it's unnecessary.